+++ contracts/modules/0_CMTATBaseGeneric.sol
@@ -14,7 +14,7 @@ import {AccessControlModule} from "./wrapper/security/AccessControlModule.sol";
 import {VersionModule} from "./wrapper/core/VersionModule.sol";
 // Extensions
 import {ExtraInformationModule} from "./wrapper/extensions/ExtraInformationModule.sol";
-import {DocumentEngineModule} from "./wrapper/extensions/DocumentEngineModule.sol";
+import {DocumentERC1643Module} from "./wrapper/extensions/DocumentERC1643Module.sol";
  /* ==== Interface and other library === */
 import {ICMTATConstructor} from "../interfaces/technical/ICMTATConstructor.sol";
 
@@ -30,7 +30,7 @@ abstract contract CMTATBaseGeneric is
     // Core
     VersionModule,
     // Extension
-    DocumentEngineModule,
+    DocumentERC1643Module,
     ExtraInformationModule,
     AccessControlModule
 {  
@@ -83,6 +83,6 @@ abstract contract CMTATBaseGeneric is
     //////////////////////////////////////////////////////////////*/
 
     /* ==== Access Control ==== */
-    function  _authorizeDocumentManagement() internal virtual override(DocumentEngineModule) onlyRole(DOCUMENT_ROLE){}
+    function  _authorizeDocumentManagement() internal virtual override(DocumentERC1643Module) onlyRole(DOCUMENT_ROLE){}
     function  _authorizeExtraInfoManagement() internal virtual override(ExtraInformationModule) onlyRole(EXTRA_INFORMATION_ROLE){}
 }