+++ contracts/modules/3_CMTATBaseRuleEngine.sol
@@ -7,7 +7,7 @@ import {CMTATBaseCommon} from "./0_CMTATBaseCommon.sol";
 import {ERC20Upgradeable} from "@openzeppelin/contracts-upgradeable/token/ERC20/ERC20Upgradeable.sol";
 /* ==== Wrapper === */
 // Core
-import {CMTATBaseAccessControl} from "./1_CMTATBaseAccessControl.sol";
+import {CMTATBaseAccessControl} from "./2_CMTATBaseAccessControl.sol";
 import {PauseModule}  from "./wrapper/core/PauseModule.sol";
 import {EnforcementModule} from "./wrapper/core/EnforcementModule.sol";
 // Extensions