+++ contracts/modules/6_CMTATBaseDebtEngine.sol
@@ -2,17 +2,18 @@
 
 pragma solidity ^0.8.20;
 
-import {ContextUpgradeable} from "@openzeppelin/contracts-upgradeable/utils/ContextUpgradeable.sol";
+/* ==== OpenZeppelin === */
 import {ERC20Upgradeable} from "@openzeppelin/contracts-upgradeable/token/ERC20/ERC20Upgradeable.sol";
-import {CMTATBaseERC2771} from "./5_CMTATBaseERC2771.sol";
-import {CMTATBaseERC20CrossChain} from "./4_CMTATBaseERC20CrossChain.sol";
+/* ==== Module === */
+import {DebtEngineModule} from "./wrapper/options/DebtEngineModule.sol";
+import {CMTATBaseERC20CrossChain} from "./5_CMTATBaseERC20CrossChain.sol";
 import {CMTATBaseSnapshot} from "./0_CMTATBaseSnapshot.sol";
 import {SnapshotEngineModule} from "./wrapper/extensions/SnapshotEngineModule.sol";
 
 /**
-* @title Extend CMTATBaseERC2771 with snapshot engine support
+* @title Extend CMTAT Base with DebtEngine module and snapshot engine support
 */
-abstract contract CMTATBaseERC2771Snapshot is CMTATBaseERC2771, CMTATBaseSnapshot {
+abstract contract CMTATBaseDebtEngine is DebtEngineModule, CMTATBaseERC20CrossChain, CMTATBaseSnapshot {
 
     /*//////////////////////////////////////////////////////////////
                   ERC-20 / CMTATBaseSnapshot disambiguation
@@ -61,30 +62,10 @@ abstract contract CMTATBaseERC2771Snapshot is CMTATBaseERC2771, CMTATBaseSnapsho
     }
 
     /*//////////////////////////////////////////////////////////////
-                  ERC2771 / Context disambiguation
-    //////////////////////////////////////////////////////////////*/
-
-    function _msgSender()
-        internal virtual view override(CMTATBaseERC2771, ContextUpgradeable) returns (address sender)
-    {
-        return CMTATBaseERC2771._msgSender();
-    }
-
-    function _msgData()
-        internal virtual view override(CMTATBaseERC2771, ContextUpgradeable) returns (bytes calldata)
-    {
-        return CMTATBaseERC2771._msgData();
-    }
-
-    function _contextSuffixLength()
-        internal virtual view override(CMTATBaseERC2771, ContextUpgradeable) returns (uint256)
-    {
-        return CMTATBaseERC2771._contextSuffixLength();
-    }
-
-    /*//////////////////////////////////////////////////////////////
                         Access Control
     //////////////////////////////////////////////////////////////*/
+
+    function _authorizeDebtEngineManagement() internal virtual override(DebtEngineModule) onlyRole(DEBT_ENGINE_ROLE) {}
 
     function _authorizeSnapshots() internal virtual override(SnapshotEngineModule) onlyRole(SNAPSHOOTER_ROLE) {}
 }