+++ contracts/modules/6_CMTATBaseERC2612.sol
@@ -7,7 +7,7 @@ import {ERC20PermitUpgradeable} from "@openzeppelin/contracts-upgradeable/token/
 import {ERC20Upgradeable} from "@openzeppelin/contracts-upgradeable/token/ERC20/ERC20Upgradeable.sol";
 import {MulticallUpgradeable} from "@openzeppelin/contracts-upgradeable/utils/MulticallUpgradeable.sol";
 /* ==== Module === */
-import {CMTATBaseERC20CrossChain, CMTATBaseRuleEngine} from "./4_CMTATBaseERC20CrossChain.sol";
+import {CMTATBaseERC20CrossChain, CMTATBaseRuleEngine} from "./5_CMTATBaseERC20CrossChain.sol";
 /* ==== Interface and other library === */
 import {ICMTATConstructor} from "../interfaces/technical/ICMTATConstructor.sol";